requirements.txt
httpx
fastmcp
inflect
